erp/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
SQLAlchemy
psycopg[binary]
jinja2
